Move commits and xrepo into shared directory.
