+++ lsif/src/test-utils.ts
@@ -6,7 +6,7 @@ import { connectPostgres } from './connection'
 import { lsp } from 'lsif-protocol'
 import { child_process } from 'mz'
 import { convertLsif } from './worker/importer/importer'
-import { XrepoDatabase } from './xrepo'
+import { XrepoDatabase } from './shared/xrepo/xrepo'
 import { dbFilename, ensureDirectory } from './util'
 import { userInfo } from 'os'
 import * as constants from './constants'
