+++ lsif/src/tests/util.ts
@@ -2,7 +2,7 @@ import rmfr from 'rmfr'
 import { Backend, ReferencePaginationCursor } from '../server/backend/backend'
 import { Configuration } from '../config'
 import { createCleanPostgresDatabase, createStorageRoot, convertTestData } from '../test-utils'
-import { XrepoDatabase } from '../xrepo'
+import { XrepoDatabase } from '../shared/xrepo/xrepo'
 import { lsp } from 'lsif-protocol'
 
 /**
