+++ lsif/src/worker/importer/importer.ts
@@ -7,7 +7,7 @@ import { Edge, MonikerKind, RangeId, Vertex } from 'lsif-protocol'
 import { EntityManager } from 'typeorm'
 import { gzipJSON } from '../../encoding'
 import { isEqual, uniqWith } from 'lodash'
-import { Package, SymbolReferences } from '../../xrepo'
+import { Package, SymbolReferences } from '../../shared/xrepo/xrepo'
 import { Readable } from 'stream'
 import { readGzippedJsonElements } from './input'
 import { TableInserter } from '../../inserter'
